fix: use discord system message content
Discord messages' contents are empty for system messages, such as greetings for new guild members, but the messages' system contents are always set to the text users see. Using the system content instead of the content as context for the friend fixes an "empty content" error when reading system messages.