test: type pinecone query result as QueryResponse
Previously, this unit test configured pinecone to return a mock. However, the unit test resulted in an attribute error, because the mock did not implement the `__getitem__` method. Instead of adding a complex implementation to the mock, the test now uses a real `QueryResponse` instance (without performing an actual query).